refactor: extract renderers module and deduplicate epilogue (#351)
