+++ .github/copilot-instructions.md
@@ -62,6 +62,7 @@ This repository provides standardized automated workflows for managing issues. A
 ## Project Structure
 - `src/breakfast/` — package source code
   - `cli.py` — Click command definition and entry point
+  - `renderers.py` — Output formatters and table fitting
   - `api.py` — GitHub API interaction logic
   - `config.py` — TOML configuration and filtering
   - `ui.py` — Terminal formatting and progress emojis
