+++ CONTRIBUTING.md
@@ -29,6 +29,7 @@ Thanks for your interest in contributing to **breakfast**! This guide covers set
 
 - `src/breakfast/` — package source code
   - `cli.py` — Click command definition and entry point
+  - `renderers.py` — Output formatters and table fitting
   - `api.py` — GitHub API interaction logic
   - `config.py` — TOML configuration and filtering
   - `ui.py` — Terminal formatting and progress emojis
