+++ tests/test_cli.py
@@ -9,7 +9,6 @@ from unittest.mock import patch
 import pytest
 import requests
 from click.testing import CliRunner
-from tabulate import tabulate
 
 from breakfast import api, cache, cli
 
@@ -1192,75 +1191,6 @@ def _make_pr_fixture(title="Test PR", number=1):
     }
 
 
-def test_table_width_matches_tabulate_border():
-    """_table_width must agree with the actual tabulate outline border width."""
-    test_cases = [
-        [{"Repo": "repo", "PR Title": "title", "Author": "alice"}],
-        [
-            {"Repo": "short", "PR Title": "a", "Author": "bob"},
-            {
-                "Repo": "a-very-long-repository-name",
-                "PR Title": "longer title here",
-                "Author": "carol",
-            },
-        ],
-        [{"Col": "x"} for _ in range(10)],
-        [{"A": "hello", "B": "\x1b[32mgreen\x1b[0m", "C": "plain"}],
-    ]
-    for rows in test_cases:
-        plain_rows = [{k: cli._strip_ansi(v) for k, v in row.items()} for row in rows]
-        expected = len(
-            tabulate(
-                plain_rows,
-                headers="keys",
-                showindex="always",
-                tablefmt="outline",
-                disable_numparse=True,
-            ).splitlines()[0]
-        )
-        assert cli._table_width(rows) == expected, f"Mismatch for rows={rows!r}"
-
-
-def test_auto_fit_measures_later_rows_when_fitting_table():
-    rows = [
-        {
-            "Repo": "short",
-            "PR Title": "short",
-            "Author": "alice",
-            "State": "open",
-            "Files": "1",
-            "Commits": "1",
-            "+/-": "+1/-0",
-            "Comments": "0",
-            "Mergeable?": "yes (clean)",
-            "Link": "PR-1",
-        },
-        {
-            "Repo": "a-very-long-repository-name-that-should-be-truncated",
-            "PR Title": "short",
-            "Author": "alice",
-            "State": "open",
-            "Files": "1",
-            "Commits": "1",
-            "+/-": "+1/-0",
-            "Comments": "0",
-            "Mergeable?": "yes (clean)",
-            "Link": "PR-2",
-        },
-    ]
-
-    terminal_width = cli._table_width(rows[:1])
-    fitted_rows = cli._auto_fit(rows, terminal_width, explicit_max_title_length=None)
-    rendered_width = len(
-        tabulate(
-            fitted_rows, headers="keys", showindex="always", tablefmt="outline"
-        ).splitlines()[0]
-    )
-
-    assert rendered_width <= terminal_width
-    assert fitted_rows[1]["Repo"].endswith("…")
-
-
 def test_cli_status_columns_use_ascii_to_keep_rows_aligned(monkeypatch):
     monkeypatch.setattr(cli, "SECRET_GITHUB_TOKEN", "token-123")
     monkeypatch.setattr(cli, "BREAKFAST_ITEMS", ["*"])
@@ -1439,53 +1369,6 @@ def test_auto_fit_compresses_mergeable_before_dropping(monkeypatch):
     assert result.exit_code == 0
     # "(clean)" reason should be gone if compression kicked in
     assert "(clean)" not in result.stdout
-
-
-def test_auto_fit_truncates_branches_before_repo():
-    """Branches should be truncated before the repo name (#175)."""
-    long_branch = "a-very-long-head-branch-name-that-should-go-first"
-    long_repo = "a-long-repo-name-that-should-survive-longer"
-    rows = [
-        {
-            "Repo": long_repo,
-            "PR Title": "Short title",
-            "Author": "alice",
-            "State": "open",
-            "Head Branch": long_branch,
-            "Base Branch": "main",
-            "Comments": "0",
-            "Link": "PR-1",
-        }
-    ]
-
-    # Width just wide enough for the full row minus the long branch — forces
-    # head-branch truncation but not necessarily repo truncation.
-    width_without_long_branch = cli._table_width(
-        [{**rows[0], "Head Branch": "short-branch"}]
-    )
-    result = cli._auto_fit(
-        rows, width_without_long_branch, explicit_max_title_length=None
-    )
-
-    visible_head = cli._strip_ansi(result[0]["Head Branch"])
-    visible_repo = cli._strip_ansi(result[0]["Repo"])
-
-    # Head branch should have been truncated (it's longer and was tried first)
-    assert visible_head != long_branch, "Head Branch should have been truncated"
-    # Repo should still be intact because truncating the branch was enough
-    assert (
-        visible_repo == long_repo
-    ), "Repo should NOT have been truncated before branches"
-
-
-def test_auto_fit_renames_mergeable_to_mrg(monkeypatch):
-    rows = [{"Mergeable?": "✅", "PR Title": "x", "Repo": "r", "Author": "a"}]
-    # Set terminal width just narrow enough to trigger step 4b but not step 5+
-    width = cli._table_width(rows) - 1
-    result = cli._auto_fit(rows, width, explicit_max_title_length=None)
-    keys = list(result[0].keys())
-    assert "Mrg" in keys
-    assert "Mergeable?" not in keys
 
 
 def test_auto_fit_drops_columns_when_very_narrow(monkeypatch):
@@ -2257,120 +2140,6 @@ def test_cli_legendary_only_implies_legendary_marking(monkeypatch):
 # ---------------------------------------------------------------------------
 
 
-def test_compress_styled_preserves_ansi_colour():
-    import click
-
-    styled = click.style("✅ pass", fg="green", bold=True)
-    compressed = cli._compress_styled(styled)
-    # Should keep the emoji but drop " pass"
-    assert "✅" in cli._strip_ansi(compressed)
-    assert "pass" not in cli._strip_ansi(compressed)
-    # ANSI colour codes should be preserved
-    assert "\x1b[" in compressed
-
-
-def test_compress_styled_noop_for_single_word():
-    import click
-
-    styled = click.style("✅", fg="green", bold=True)
-    assert cli._compress_styled(styled) == styled
-
-
-def test_compress_styled_plain_text():
-    assert cli._compress_styled("hello world") == "hello"
-    assert cli._compress_styled("single") == "single"
-
-
-def test_compress_styled_preserves_approval_fraction():
-    styled = cli.format_approval_status(
-        "pending",
-        current_reviews=1,
-        required_reviews=2,
-    )
-
-    compressed = cli._compress_styled(styled)
-
-    assert cli._strip_ansi(compressed) == "✅ 1/2"
-
-
-def test_truncate_formatted_text_preserves_osc8_anchor():
-    linked_repo = cli.generate_terminal_url_anchor(
-        "https://github.com/myorg/really-long-repo-name",
-        "really-long-repo-name",
-    )
-
-    truncated = cli._truncate_formatted_text(linked_repo, 8)
-
-    assert cli._strip_ansi(truncated) == "really-…"
-    assert "\x1b]8;;https://github.com/myorg/really-long-repo-name\x1b\\" in truncated
-    assert truncated.endswith("\x1b]8;;\x1b\\")
-    assert "]8;;ht…" not in truncated
-
-
-def test_truncate_col_preserves_repo_and_author_hyperlinks():
-    rows = [
-        {
-            "Repo": cli.generate_terminal_url_anchor(
-                "https://github.com/myorg/really-long-repo-name",
-                "really-long-repo-name",
-            ),
-            "PR Title": "Short title",
-            "Author": cli.generate_terminal_url_anchor(
-                "https://github.com/some-very-long-author-name",
-                "some-very-long-author-name",
-            ),
-            "State": "open",
-            "Files": "1",
-            "Commits": "1",
-            "+/-": "+1/-0",
-            "Comments": "0",
-            "Mergeable?": "✅ (clean)",
-            "Link": "PR-1",
-        }
-    ]
-
-    truncated = cli._truncate_col(rows, "Repo", terminal_width=40, min_len=8)
-    truncated = cli._truncate_col(truncated, "Author", terminal_width=40, min_len=8)
-
-    assert (
-        "\x1b]8;;https://github.com/myorg/really-long-repo-name\x1b\\"
-        in truncated[0]["Repo"]
-    )
-    assert (
-        "\x1b]8;;https://github.com/some-very-long-author-name\x1b\\"
-        in truncated[0]["Author"]
-    )
-    assert "]8;;ht…" not in truncated[0]["Repo"]
-    assert "]8;;ht…" not in truncated[0]["Author"]
-
-
-def test_auto_fit_preserves_checks_colour(monkeypatch):
-    import click
-
-    styled_checks = click.style("✅ pass", fg="green", bold=True)
-    rows = [
-        {
-            "Repo": "myrepo",
-            "PR Title": "Some title",
-            "Author": "alice",
-            "State": "open",
-            "Files": "1",
-            "Commits": "1",
-            "+/-": "+1/-0",
-            "Comments": "0",
-            "Checks": styled_checks,
-            "Mergeable?": click.style("✅ (clean)", fg="green", bold=True),
-            "Link": "PR-1",
-        }
-    ]
-    # Very narrow width to force all compression steps
-    result = cli._auto_fit(rows, 80, explicit_max_title_length=None)
-    checks_key = "Checks" if "Checks" in result[0] else None
-    if checks_key:
-        # Colour should be preserved even after compression
-        assert "\x1b[" in result[0][checks_key]
-
-
 def test_no_drafts_and_drafts_only_are_mutually_exclusive(monkeypatch):
     monkeypatch.setattr(cli, "SECRET_GITHUB_TOKEN", "token-123")
 
@@ -2379,19 +2148,6 @@ def test_no_drafts_and_drafts_only_are_mutually_exclusive(monkeypatch):
 
     assert result.exit_code == 1
     assert "mutually exclusive" in result.stderr.lower()
-
-
-def test_styled_hyperlink_puts_colour_outside_osc8():
-    import click
-
-    styled = click.style("pending", fg="yellow", bold=True)
-    result = cli._styled_hyperlink("https://example.com/checks", styled)
-    # Find the link text between the OSC 8 open and close tags
-    osc_open_end = result.index("\x1b\\") + 2
-    osc_close_start = result.index("\x1b]8;;\x1b\\", osc_open_end)
-    link_text = result[osc_open_end:osc_close_start]
-    assert link_text == "pending"  # plain text inside the OSC 8, no escape sequences
-    assert "\x1b[" in result  # colour codes still present outside the OSC 8
 
 
 def test_cli_repo_and_author_are_hyperlinks(monkeypatch):
@@ -4014,41 +3770,6 @@ template = "{repo} -> {title}"
 
     assert result.exit_code == 0
     assert "repo -> My PR" in result.stdout
-
-
-def test_table_width_auto_fit_emoji_regression():
-    """Verify that _table_width accounts for double-width emojis correctly
-
-    so that auto-fit does not underestimate table display width.
-    """
-    rows = [
-        {
-            "Repo": "short",
-            "PR Title": "short title",
-            "Author": "alice",
-            "Checks": "✅ pass",
-            "Approved": "⏳ pending",
-            "Mergeable?": "❌ (dirty)",
-        }
-    ]
-    # Estimate width using _table_width
-    estimated_width = cli._table_width(rows)
-
-    # Actual width from tabulate (with wcwidth installed)
-    from tabulate import tabulate
-
-    actual_width = len(
-        tabulate(
-            rows,
-            headers="keys",
-            showindex="always",
-            tablefmt="outline",
-            disable_numparse=True,
-        ).splitlines()[0]
-    )
-
-    # They should match exactly, preventing line wrap double spacing
-    assert estimated_width == actual_width
 
 
 _COLOUR_INDEX_PR = {
